requirements.txt
Django==3.2.5
djangorestframework==3.12.4
mysqlclient==2.0.3
drf-yasg==1.20.0
djangorestframework-simplejwt==4.7.2
google-api-python-client==2.14.0
facebook-sdk==3.1.0
